requirements.txt
streamlit==1.20.0
pandas==2.0.3
pyyaml==6.0
joblib==1.3.2
matplotlib==3.8.0
plotly==5.15.0
scikit-learn==1.3.2
